Develop

+++ src/database/migrations/2024_05_25_000000_create_users_table.php
@@ -16,7 +16,7 @@ class CreateUsersTable extends Migration
         Schema::create('users', function (Blueprint $table) {
             $table->id();
             $table->foreignId('role_id')->constrained()->cascadeOnDelete();
-            $table->unsignedTinyInteger('shop_id')->nullable();
+            $table->unsignedBigInteger('shop_id')->nullable();
             $table->string('name', 50);
             $table->string('email', 100)->unique();
             $table->timestamp('email_verified_at')->nullable();
